TST: Upgrade pytest to latest.
Rerunning the tests with the upgraded pytest produces the same number of passing
and xfailed tests.

With pytest 3.2.3, 5 warnings are now shown.
pytest 3.1.0 changed behavior so that warnings are captured and displayed.
https://docs.pytest.org/en/latest/changelog.html#id81

The impetus for upgrading is that in another patch, using `pytest.param` to mark a
specific parameter as an xfail would have been useful; `pytest.param` was added
in 3.1.0